Revert back ruby and repo-small-badge
For various reason, can't update the ruby version of a project I'm using
this on.

Revert the minimum ruby version to 2.5.0, the same required by the
latest SimpleCov, and the repo-small-badge version to 0.2.9 (which also
requires 2.5.0 as a minimum ruby version).

Hopefully this will become useless as soon as I can upgrade ruby on the
relevant project.